% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_differences.R
\name{multi_CI}
\alias{multi_CI}
\title{Multidimensional Credible Interval}
\usage{
multi_CI(data, mean, cov, df, level = 0.95)
}
\arguments{
\item{data}{A vector, of compatible dimension with \code{mean} and \code{cov}}

\item{mean}{A vector, the mean parameter of the multivariate t-distribution}

\item{cov}{A matrix, the covariance parameter of the multivariate
t-distribution}

\item{df}{A number, the degrees of freedom of the multivariate t-distribution}

\item{level}{A number, between 0 and 1, corresponding to the level of the
Credible Interval. Default is 0.95.}
}
\value{
A boolean, indicating whether the \code{data} vector belongs to the
computed Credible Interval.
}
\description{
Compute whether a given vector belongs to the multivariate t-distribution
Credible Interval for given level, mean, covariance and degrees of freedom.
}
\examples{
TRUE
}
