% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_differences.R
\name{multi_identify_diff}
\alias{multi_identify_diff}
\title{Identify differences in multivariate posteriors}
\usage{
multi_identify_diff(
  posterior,
  plot = TRUE,
  overlap_coef = TRUE,
  cumulative = FALSE,
  nb_sample = 1e+05,
  nb_sample_overlap = 10000
)
}
\arguments{
\item{posterior}{A tibble, typically coming from a \code{posterior_mean()}
function, containing the parameters of the multivariate posterior
t-distributions for the mean of the considered groups and draws for each
peptide.}

\item{plot}{A boolean, indicating whether a results plot should be displayed.}

\item{overlap_coef}{A boolean, indicating whether the overlapping coefficient
between multivariate t-distributions should be computed for all groups.}

\item{cumulative}{A boolean, indicating whether the probability distribution
should be the cumulative instead.}

\item{nb_sample}{A number of samples to draw from the empirical distributions}

\item{nb_sample_overlap}{A number of samples to draw for the Monte Carlo
approximation of the Overlapping Coefficients between all posteriors.}
}
\value{
A list, containing:
\itemize{
\item Diff_mean, a tibble containing the posterior mean (and their difference)
for each peptide and for all one-by-one group comparisons.
\item Diff_proba, a tibble containing the probability distribution of the
number of differential peptides for all one-by-one group comparisons.
\item Overlap_coef, a tibble containing the overlapping coefficient between
the posterior multivariate t-distributions for all one-by-one group
comparisons.
}
}
\description{
Compute a multivariate inference criterion to examine whether the posterior
multivariate t-distributions of groups should be considered different enough
to be called 'differential'. Two groups are considered can be discriminated
based on the probability weights for the element-wise means to be greater in
each group. The Credible Intervals for each marginals are also provided.
}
\examples{
TRUE
}
