% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_differences.R
\name{overlap_coef}
\alias{overlap_coef}
\title{Overlapping coefficient between univariate t-distributions}
\usage{
overlap_coef(mean1, mean2, var1, var2, df1, df2)
}
\arguments{
\item{mean1}{A vector, the mean parameter of a t-distribution}

\item{mean2}{A vector, the mean parameter of the other t-distribution}

\item{var1}{A matrix, the variance parameter of a t-distribution}

\item{var2}{A matrix, the variance parameter of the other
t-distribution}

\item{df1}{A number, the degrees of freedom of a t-distribution}

\item{df2}{A number, the degrees of freedom of the other t-distribution}

\item{nb_sample}{A number of samples drawn to compute the Monte Carlo estimation}
}
\value{
A number, the Monte Carlo approximation of the overlapping
coefficient between the two univariate t-distributions.
}
\description{
Compute a (high speed) quadrature approximation of the overlapping coefficient
between two univariate t-distributions with arbitrary mean, covariance and
degrees of freedom.
}
\examples{
TRUE
}
