% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posteriors.R
\name{post_mean_diff}
\alias{post_mean_diff}
\title{Title}
\usage{
post_mean_diff(data, mu_0, lambda_0, Sigma_0, nu_0)
}
\arguments{
\item{data}{A tibble or data frame containing imputed data sets for all
groups. Required columns: \code{Peptide}, \code{Output}, \code{Group},
\code{Draw}.}

\item{mu_0}{A vector, corresponding to the prior mean.}

\item{lambda_0}{A number, corresponding to the prior covariance scaling
parameter.}

\item{Sigma_0}{A matrix, corresponding to the prior covariance parameter.}

\item{nu_0}{A number, corresponding to the prior degrees of freedom.}
}
\value{
A vector providing the empirical posterior distribution for the
means of the desired groups.
}
\description{
Title
}
\examples{
TRUE
}
