% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posteriors.R
\name{post_mean_diff_uni}
\alias{post_mean_diff_uni}
\title{Title}
\usage{
post_mean_diff_uni(data, mu_0, lambda_0, beta_0, alpha_0)
}
\arguments{
\item{data}{A tibble or data frame containing imputed data sets for all
groups. Required columns: Peptide, Output, Group, Draw}

\item{mu_0}{A vector, corresponding to the prior mean.}

\item{lambda_0}{A number, corresponding to the prior covariance scaling
parameter.}

\item{beta_0}{A matrix, corresponding to the prior covariance parameter.}

\item{alpha_0}{A number, corresponding to the prior degrees of freedom.}
}
\value{
A vector providing the empirical posterior distribution of the
}
\description{
Title
}
\examples{
TRUE
}
