% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posteriors.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Posterior distribution of the means}
\usage{
posterior_mean(data, mu_0 = NULL, lambda_0 = 1, beta_0 = 1, alpha_0 = 1)
}
\arguments{
\item{data}{A tibble or data frame containing imputed data sets for all
groups. Required columns:  \code{Peptide}, \code{Output}, \code{Group},
\code{Sample}.}

\item{mu_0}{A vector, corresponding to the prior mean.}

\item{lambda_0}{A number, corresponding to the prior covariance scaling
parameter.}

\item{beta_0}{A matrix, corresponding to the prior covariance parameter.}

\item{alpha_0}{A number, corresponding to the prior degrees of freedom.}
}
\value{
A tibble providing the empirical posterior distribution for the
}
\description{
Compute the posterior distribution of the means between multiple groups.
All peptides are considered independent from one another.
}
\examples{
TRUE
}
