% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_differences.R
\name{multi_identify_diff}
\alias{multi_identify_diff}
\title{Identify differences in multivariate posteriors}
\usage{
multi_identify_diff(
  posterior,
  plot = TRUE,
  cumulative = FALSE,
  CI_level = 0.05,
  nb_sample = 10000,
  nb_sample_per_dim = NULL
)
}
\arguments{
\item{posterior}{A tibble, typically coming from a \code{posterior_mean()}
function, containing the parameters of the multivariate posterior
t-distributions for the mean of the considered groups and draws for each
peptide.}

\item{plot}{A boolean, indicating whether a results plot should be displayed.}

\item{cumulative}{A boolean, indicating whether the probability distribution
should be the cumulative instead.}

\item{CI_level}{A number, defining the order of quantile chosen to assess
differences between groups.}

\item{nb_sample}{A number of samples to draw from the empirical distributions}

\item{nb_sample_per_dim}{}
}
\value{
A tibble, indicating the probability distribution for the number of
differentiable peptides for all one-by-one group comparisons
}
\description{
Compute a multivariate inference criterion to examine whether the posterior
multivariate t-distributions of groups should be considered different enough
to be called 'differential'. Two groups are considered can be discriminated
based on the probability weights for the element-wise means to be greater in
each group. The Credible Intervals for each marginals are also provided.
}
\examples{
TRUE
}
